requirements.txt
Django>=5.1.7
channels>=4.0.0
daphne>=4.0.0
django-bootstrap5>=23.3
Pillow>=10.0.0
pdf2image>=1.16.3
mysqlclient>=2.1.0 